Web: pagesフォルダ削除
