+++ apps/web/src/app/page.tsx
@@ -1,5 +1,34 @@
-import { Page } from "@/pages";
+"use client";
+
+import { SpotCard } from "@/features/spots/SpotCard";
+import { SpotList, SpotListInnerProps } from "@/features/spots/SpotList";
+import { SpotsSearchBox } from "@/features/spots/SpotsSearchBox";
+import { Link } from "@/components/ui/link";
+import React from "react";
+
+const SpotListInner: React.FC<SpotListInnerProps> = ({ data }) => {
+  return (
+    <ul>
+      {data?.spots.data.map((value) => {
+        return (
+          <li key={value.id}>
+            <Link href={`/spots/${value.id}`}>
+              <SpotCard data={value} />
+            </Link>
+          </li>
+        );
+      })}
+    </ul>
+  );
+};
 
 export default function Home() {
-  return <Page />;
+  return (
+    <article>
+      <SpotsSearchBox />
+      <SpotList>
+        <SpotListInner />
+      </SpotList>
+    </article>
+  );
 }
