--- apps/web/src/pages/courses/originals/create/index.tsx
@@ -1,31 +1,0 @@
-"use client";
-
-import { SpotCard } from "@/features/spots/SpotCard";
-import { SpotList, SpotListInnerProps } from "@/features/spots/SpotList";
-import { SpotsSearchBox } from "@/features/spots/SpotsSearchBox";
-import React from "react";
-
-const SpotListInner: React.FC<SpotListInnerProps> = ({ data }) => {
-  return (
-    <ul>
-      {data?.spots.data.map((value) => {
-        return (
-          <li key={value.id}>
-            <SpotCard data={value} />
-          </li>
-        );
-      })}
-    </ul>
-  );
-};
-
-export function Page() {
-  return (
-    <article>
-      <SpotsSearchBox />
-      <SpotList>
-        <SpotListInner />
-      </SpotList>
-    </article>
-  );
-}
